Make the code compile and work
There are several reasons why one may be not fully satisfied with
the current code, so I anticipate some more hammering is required before
this can be called ready as a PR.

Cc: @lorenzoPrimi (this work and PR is part of the effort to enable
you to smoothly transition from old to new APIs w/ iOS).

+++ src/measurement_kit/cmd/multi_ndt.cpp
@@ -23,7 +23,7 @@ int main(std::list<Callback<BaseTest &>> &initializers, int argc, char **argv) {
         case 'u':
             // By default only the download phase is performed
             test.set_option("single_test_suite",
-                    MK_NDT_DOWNLOAD | MK_NDT_UPLOAD);
+                    (int64_t)(MK_NDT_DOWNLOAD | MK_NDT_UPLOAD));
             break;
         default:
             fprintf(stderr, "%s\n", USAGE);
